Move state api request building to collaborator
